smt error refactoring
